BACK-CORE: response time second to millisec

+++ domain/test.go
@@ -13,7 +13,7 @@ type Test struct {
 	Method         string             `bson:"method" json:"method" validate:"required"`
 	ResponseStatus int                `bson:"response_status" json:"response_status" validate:"required"`
 	ResponseBdy    string             `bson:"response_bdy" json:"response_bdy" validate:"required"`
-	ResponseTime   float64            `bson:"response_time" json:"response_time" validate:"required"`
+	ResponseTime   int64              `bson:"response_time" json:"response_time" validate:"required"`
 	TestID         string             `bson:"test_id" json:"test_id" validate:"required"`
 	CreatedAt      time.Time          `bson:"created_at" json:"created_at"`
 	UpdatedAt      time.Time          `bson:"updated_at,omitempty" json:"updated_at"`
